+++ tests/integration/cqlengine/statements/test_base_statement.py
@@ -138,10 +138,10 @@ class ExecuteStatementTest(BaseCassEngTestCase):
         self.assertEqual(result.one()["text"], self.text)
 
         q = TestQueryUpdateModel.objects.filter(text__like=like_clause).allow_filtering()
-        self.assertEqual(q[0].text, self.text)
+        self.assertEqual(q.first().text, self.text)
 
         q = TestQueryUpdateModel.objects.filter(text__like=like_clause)
-        self.assertEqual(q[0].text, self.text)
+        self.assertEqual(q.first().text, self.text)
 
     def _insert_statement(self, partition, cluster):
         # Verifying insert statement